Se extrajo la lista con boton para remover en un componente
